Switched floats to ints in gradient shaders for better performance (flutter/engine#40189)
Switched floats to ints in gradient shaders for better performance

+++ engine/src/flutter/impeller/entity/shaders/radial_gradient_ssbo_fill.frag
@@ -21,7 +21,7 @@ uniform FragInfo {
   float radius;
   float tile_mode;
   float alpha;
-  float colors_length;
+  int colors_length;
 }
 frag_info;
